% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CITEseq_example}
\alias{CITEseq_example}
\title{A subset of ECCITE-seq data (control)}
\format{
An object of class \code{list} of length 3.
}
\source{
Gene Expression Omnibus with the accession code GSE126310.
}
\usage{
data(CITEseq_example, package = 'CiteFuse')
}
\description{
Data from Mimitou et al. ECCITE-seq PBMC control sample data, which is a list
of three matrices of RNA, ADT and HTO
}
\references{
Mimitou, E. P., Cheng, A., Montalbano, A., et al. (2019).
Multiplexed detection of proteins, transcriptomes, clonotypes and
CRISPR perturbations in single cells.
Nature Methods, 16(5), 409–412.
}
\keyword{datasets}
