% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{eigenMapMatMult}
\alias{eigenMapMatMult}
\title{Rcpp function for matrix}
\usage{
eigenMapMatMult(A, B, C)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix}

\item{C}{a matrix}
}
\value{
A matrix
}
\description{
Rcpp function for matrix
}
