% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importanceADT.R
\name{importanceADT}
\alias{importanceADT}
\title{A function to calculate the importance score of ADT}
\usage{
importanceADT(sce, altExp_name = "ADT", exprs_value = "logcounts",
  group = NULL, subsample = TRUE, times = 10, prop = 0.8, ...)
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{altExp_name}{A character indicates which expression matrix is used. by default is none (i.e. RNA).}

\item{exprs_value}{A character indicates which expression value in assayNames is used.}

\item{group}{A vector indicates the grouping of the data}

\item{subsample}{Whether perform subsampling}

\item{times}{A numeric indicates the times of subsampling is performed}

\item{prop}{A numeric indicates the proportion of cells are subsampled from the whole data}

\item{...}{other arguments to `randomForest()` function}
}
\description{
A function to calculate the importance score of ADT
}
