% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligandReceptor_analysis.R
\name{ligandReceptorTest}
\alias{ligandReceptorTest}
\title{A function to perform ligand receptor analysis}
\usage{
ligandReceptorTest(sce, ligandReceptor_list, cluster,
  RNA_exprs_value = "minMax", altExp_name = "ADT",
  altExp_exprs_value = "zi_minMax", num_permute = 1000, p_sig = 0.05,
  ncores = 1)
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{ligandReceptor_list}{A data.frame indicates the ligand receptor list}

\item{cluster}{A vector indicates the cluster results}

\item{RNA_exprs_value}{A character indicates which expression value for RNA in assayNames is used.}

\item{altExp_name}{A character indicates which expression matrix is used. by default is none (i.e. RNA).}

\item{altExp_exprs_value}{A character indicates which expression value in assayNames is used.}

\item{num_permute}{Number of permutation.}

\item{p_sig}{A numeric indicates threshold of the pvalue significance}

\item{ncores}{Number of cores can be used}
}
\description{
A function to perform ligand receptor analysis
}
