% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{louvain}
\alias{louvain}
\title{A function to perform louvain clustering}
\usage{
louvain(sce, metadata = "SNF_W")
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{metadata}{indicates the meta data name of affinity matrix to virsualise}
}
\description{
A function to perform louvain clustering
}
