% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfunctions.R
\name{plotHTO}
\alias{plotHTO}
\title{A function to plot HTO expression}
\usage{
plotHTO(sce, which_idx = 1:2, altExp_name = NULL, ncol = 2)
}
\arguments{
\item{sce}{sce}

\item{which_idx}{which_idx}

\item{altExp_name}{altExp_name}

\item{ncol}{ncol}
}
\description{
A function to plot HTO expression
}
