% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfunctions.R
\name{plotHTOSingle}
\alias{plotHTOSingle}
\title{A function to plot HTO expression}
\usage{
plotHTOSingle(sce, which_idx = 1:2, altExp_name = NULL)
}
\arguments{
\item{sce}{sce}

\item{which_idx}{which_idx}

\item{altExp_name}{altExp_name}
}
\description{
A function to plot HTO expression
}
