% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfunctions.R
\name{readFrom10X}
\alias{readFrom10X}
\title{A function to read the data from 10X}
\usage{
readFrom10X(
  dir,
  type = c("auto", "sparse", "HDF5"),
  feature_named_by = c("gene_id", "gene_symbol"),
  filter_features = TRUE
)
}
\arguments{
\item{dir}{A character indicates the directory of the 10X files}

\item{type}{A character indicates the format of the data, sparse or HDF5}

\item{feature_named_by}{A character indicates whehter the genes will be named by gene_id or gene_symbol}

\item{filter_features}{A logical input indicates whether the features with all zeros will be removed}
}
\description{
A function to read the data from 10X
}
