% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{reducedDimSNF}
\alias{reducedDimSNF}
\title{reducedDimSNF}
\usage{
reducedDimSNF(sce, metadata = "SNF_W", method = "UMAP", dimNames = NULL, ...)
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{metadata}{indicates the meta data name of
affinity matrix to virsualise}

\item{method}{the method of visualisation, which can be UMAP,
tSNE and diffusion map}

\item{dimNames}{indicates the name of the reduced dimension results.}

\item{...}{other parameters for tsne(), umap()}
}
\value{
A SingleCellExperiment object
}
\description{
A function to reduce the dimension of the similarity matrix
}
\examples{
data(sce_control_subset)
sce_control_subset <- CiteFuse(sce_control_subset)
sce_control_subset <- reducedDimSNF(sce_control_subset,
method = "tSNE",
dimNames = "tSNE_joint")

}
