% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sce_control_subset}
\alias{sce_control_subset}
\title{A SingleCellExperiment of ECCITE-seq data}
\format{
An object of class \code{SingleCellExperiment} with 1508 rows and 128 columns.
}
\source{
Gene Expression Omnibus with the accession code GSE126310.
}
\usage{
data(sce_control_subset, package = 'CiteFuse')
}
\description{
Data from Mimitou et al. ECCITE-seq PBMC Control sample data
}
\references{
Mimitou, E. P., Cheng, A., Montalbano, A., et al. (2019).
Multiplexed detection of proteins, transcriptomes, clonotypes and
CRISPR perturbations in single cells.
Nature Methods, 16(5), 409–412.
}
\keyword{datasets}
