% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNF_fast.R
\name{SNF_fast}
\alias{SNF_fast}
\title{Fast similarity network fusion method}
\usage{
SNF_fast(Wall, K = 20, t = 20)
}
\arguments{
\item{Wall}{List of matrices. Each element of the list is a square, symmetric matrix that shows affinities of the data points from a certain view.}

\item{K}{Number of neighbors in K-nearest neighbors part of the algorithm.}

\item{t}{Number of iterations for the diffusion process.}
}
\value{
W is the overall status matrix derived
}
\description{
Fast similarity network fusion method
}
