% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{spectralClustering}
\alias{spectralClustering}
\title{spectralClustering}
\usage{
spectralClustering(
  affinity,
  K = 20,
  type = 4,
  fast = TRUE,
  maxdim = 50,
  delta = 1e-05,
  t = 0,
  neigen = NULL
)
}
\arguments{
\item{affinity}{An affinity matrix}

\item{K}{number of clusters}

\item{type}{type}

\item{fast}{fast}

\item{maxdim}{maxdim}

\item{delta}{delta}

\item{t}{t}

\item{neigen}{neigen}
}
\value{
A list indicates the spectral clustering results
}
\description{
A function to perform spectral clustering
}
\examples{

data("sce_control_subset", package = "CiteFuse")
sce_control_subset <- CiteFuse(sce_control_subset)
SNF_W_clust <- spectralClustering(S4Vectors::metadata(sce_control_subset)[["SNF_W"]],
K = 5)

}
