% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{visualiseDim}
\alias{visualiseDim}
\title{A function to visualise the reduced dimension}
\usage{
visualiseDim(
  sce,
  dimNames = NULL,
  colour_by = NULL,
  shape_by = NULL,
  dim = 1:2
)
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{dimNames}{indicates the name of the reduced dimension results.}

\item{colour_by}{the method of visualisation, which can be UMAP, tSNE and diffusion map}

\item{shape_by}{the method of visualisation, which can be UMAP, tSNE and diffusion map}

\item{dim}{a vector of numeric with length of 2 indicates which component is being plot}
}
\description{
A function to visualise the reduced dimension
}
