% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{visualiseKNN}
\alias{visualiseKNN}
\title{visualiseKNN}
\usage{
visualiseKNN(sce, colour_by = NULL, metadata = "SNF_W")
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{colour_by}{the name of coldata that is used to colour the node}

\item{metadata}{indicates the meta data name of affinity matrix to virsualise}
}
\value{
A igraph plot
}
\description{
A function to perform louvain clustering
}
