% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfunctions.R
\name{crossSampleDoublets}
\alias{crossSampleDoublets}
\title{A function that perform normalisation for alternative expression}
\usage{
crossSampleDoublets(sce, altExp_name = NULL, totalExp_threshold = 10)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object}

\item{altExp_name}{Name of alternative expression that will be used to perform normalisation.
If it is NULL, it will set to HTO.}

\item{totalExp_threshold}{the threshold indicates for the HTO less than this threshold
will be filtered from the analysis}
}
\description{
A function that perform normalisation for alternative expression
}
