% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lr_pair_subset}
\alias{lr_pair_subset}
\title{A subset of Ligand Receptor Pairs}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 50 rows and 2 columns.
}
\usage{
lr_pair_subset
}
\description{
A subset of Ligand Receptor Pairs
}
\keyword{datasets}
