% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgenes.R
\name{selectDEgenes}
\alias{selectDEgenes}
\title{A function to select DE genes}
\usage{
selectDEgenes(de_res, pval_adj = 0.05, mean_diff = 0, pct_diff = 0.1,
  topN = 10)
}
\arguments{
\item{de_res}{The de results returned by `DEgenes()``}

\item{pval_adj}{A numeric indicates the threshold of adjusted p-value.}

\item{mean_diff}{A numeric indicates the threshold of difference of average expression.}

\item{pct_diff}{A numeric indicates the threshold of difference of percentage expression.}

\item{topN}{A numeric indicates the top number of genes will be included in the list.}
}
\description{
A function to select DE genes
}
