% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{spectralClustering}
\alias{spectralClustering}
\title{A function to perform spectral clustering}
\usage{
spectralClustering(
  affinity,
  K = 20,
  type = 4,
  fast = T,
  maxdim = 50,
  delta = 1e-05,
  t = 0,
  neigen = NULL
)
}
\arguments{
\item{affinity}{An affinity matrix}

\item{K}{number of clusters}

\item{type}{type}

\item{fast}{fast}

\item{maxdim}{maxdim}

\item{delta}{delta}

\item{t}{t}

\item{neigen}{neigen}
}
\description{
A function to perform spectral clustering
}
