% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ligandReceptor_analysis.R
\name{visLigandReceptor}
\alias{visLigandReceptor}
\title{A function to visualise ligand receptor analysis}
\usage{
visLigandReceptor(sce, type = c("pval_heatmap", "pval_dotplot",
  "group_network", "group_heatmap", "lr_network"))
}
\arguments{
\item{sce}{A singlecellexperiment object}

\item{type}{A character indicates the type of the plot}
}
\description{
A function to visualise ligand receptor analysis
}
