% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QCfunctions.R
\name{withinSampleDoublets}
\alias{withinSampleDoublets}
\title{doublet identification within batch}
\usage{
withinSampleDoublets(sce, altExp_name = NULL, eps = 200, minPts = 50)
}
\arguments{
\item{sce}{a SingleCellExperiment}

\item{altExp_name}{expression name of HTO matrix}

\item{eps}{eps of DBSCAN}

\item{minPts}{minPts of DBSCAN}
}
\description{
doublet identification within batch
}
